Moved tooltip to extensions.
