+++ tools/gulptasks/jsdoc-classes.js
@@ -38,7 +38,6 @@ const SOURCE_GLOBS = [
     'js/Core/Globals.js',
     'js/Core/Pointer.js',
     'js/Core/Time.js',
-    'js/Core/Tooltip.js',
     'js/Core/Utilities.js',
     'js/Data/DataTable.js',
     'js/Extensions/Annotations/Controllables/Controllable.js',
@@ -61,6 +60,7 @@ const SOURCE_GLOBS = [
     'js/Extensions/OfflineExporting/OfflineExporting.js',
     'js/Extensions/PatternFill.js',
     'js/Extensions/SeriesLabel/SeriesLabel.js',
+    'js/Extensions/Tooltip/Tooltip.js',
     'js/Maps/MapNavigation.js',
     'js/Maps/MapView.js',
     'js/Series/AreaRange/AreaRangeSeries.js',
