+++ ts/Core/Pointer.ts
@@ -37,7 +37,7 @@ const {
     noop
 } = H;
 import { Palette } from '../Core/Color/Palettes.js';
-import Tooltip from './Tooltip.js';
+import Tooltip from '../Extensions/Tooltip/Tooltip.js';
 import U from './Utilities.js';
 import SVGAttributes from './Renderer/SVG/SVGAttributes';
 import BBoxObject from './Renderer/BBoxObject';
