+++ ts/Core/Series/Series.ts
@@ -47,7 +47,7 @@ import type { StatesOptionsKey } from './StatesOptions';
 import type SVGAttributes from '../Renderer/SVG/SVGAttributes';
 import type SVGPath from '../Renderer/SVG/SVGPath';
 import type { SymbolKey } from '../Renderer/SVG/SymbolType';
-import type TooltipOptions from '../TooltipOptions';
+import type TooltipOptions from '../../Extensions/Tooltip/TooltipOptions';
 
 import A from '../Animation/AnimationUtilities.js';
 const {
