+++ ts/Extensions/Annotations/Controllables/Controllable.ts
@@ -24,7 +24,7 @@ import type SVGElement from '../../../Core/Renderer/SVG/SVGElement';
 
 import ControlPoint from '../ControlPoint.js';
 import MockPoint from '../MockPoint.js';
-import Tooltip from '../../../Core/Tooltip.js';
+import Tooltip from '../../Tooltip/Tooltip.js';
 import U from '../../../Core/Utilities.js';
 const {
     isObject,
