+++ ts/Stock/Indicators/OBV/OBVOptions.d.ts
@@ -17,7 +17,6 @@ import type {
     SMAParamsOptions
 } from '../SMA/SMAOptions';
 import type { PointMarkerOptions } from '../../../Core/Series/PointOptions';
-import type TooltipOptions from '../../../Core/TooltipOptions';
 
 /* *
  *
@@ -34,5 +33,11 @@ export interface OBVParamsOptions extends SMAParamsOptions {
     // for inheritance
     volumeSeriesID?: string;
 }
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
 
 export default OBVOptions;
